requirements.txt
openai==0.25.0
PyGithub==1.57
